Refactor backend code and update submission service
