+++ speed-backend/src/main.ts
@@ -1,10 +1,11 @@
 import { NestFactory } from '@nestjs/core';
 import { AppModule } from './app.module';
-import * as cors from 'cors';
+import { ConfigModule } from '@nestjs/config';
 
 async function bootstrap() {
+  ConfigModule.forRoot() // Load environment variables
   const app = await NestFactory.create(AppModule);
-  app.use(cors());
-  await app.listen(3001);
+  app.enableCors();
+  await app.listen(process.env.PORT, () => console.log(`\nApp listen at http://localhost:${process.env.PORT}/`));
 }
 bootstrap();
