+++ speed-backend/src/submissions/interfaces/submission.interface.ts
@@ -1,0 +1,9 @@
+export interface Submission extends Document {
+    title: string;
+    authors: string;
+    journal: string;
+    year: string;
+    volume?: string;
+    pages?: string;
+    doi: string;
+}
